requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14,<4.0
gunicorn>=21.0,<22.0
psycopg2-binary>=2.9,<3.0  # PostgreSQL desteği için
dj-database-url>=1.0,<2.0  # Veritabanı URL yönetimi için
python-decouple>=3.8,<4.0  # Ortam değişkenleri için